+++ paddle/fluid/operators/activation_op.cu
@@ -11,9 +11,9 @@ limitations under the License. */
 
 #include "paddle/fluid/operators/activation_op.h"
 #include "paddle/fluid/operators/amp/fp16_type_traits.h"
-#include "paddle/fluid/operators/elementwise/elementwise_op_impl.cu.h"
 #include "paddle/fluid/platform/bfloat16.h"
 #include "paddle/fluid/platform/device/gpu/gpu_device_function.h"
+#include "phi/kernels/funcs/elementwise_op_impl.cu.h"
 
 namespace paddle {
 namespace operators {
@@ -1367,14 +1367,14 @@ class ELUGradCudaKernel : public framework::OpKernel<T> {
     if (alpha > 0) {
       CudaELUGradFunctor<T> functor;
       functor.alpha = alpha;
-      paddle::operators::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins,
-                                                                &outs, functor);
+      phi::funcs::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins, &outs,
+                                                         functor);
     } else {
       CudaELUGradNegativeAlphaFunctor<T> functor;
       functor.alpha = alpha;
       ins.push_back(x);
-      paddle::operators::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins,
-                                                                &outs, functor);
+      phi::funcs::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins, &outs,
+                                                         functor);
     }
   }
 };
@@ -1450,8 +1450,8 @@ class ActivationCudaKernel
     for (auto& attr : attrs) {
       *attr.second = ctx.Attr<float>(attr.first);
     }
-    paddle::operators::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins,
-                                                              &outs, functor);
+    phi::funcs::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins, &outs,
+                                                       functor);
   }
 };
 
@@ -1480,17 +1480,17 @@ class ActivationGradCudaKernel
     if (static_cast<int>(Functor::FwdDeps()) == static_cast<int>(kDepOut)) {
       // Only need forward output Out
       ins.push_back(out);
-      paddle::operators::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins,
-                                                                &outs, functor);
+      phi::funcs::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins, &outs,
+                                                         functor);
     } else if (static_cast<int>(Functor::FwdDeps()) ==
                static_cast<int>(kDepX)) {
       // Only need forward input X
       ins.push_back(x);
-      paddle::operators::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins,
-                                                                &outs, functor);
+      phi::funcs::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins, &outs,
+                                                         functor);
     } else {
-      paddle::operators::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins,
-                                                                &outs, functor);
+      phi::funcs::LaunchSameDimsElementwiseCudaKernel<T>(dev_ctx, ins, &outs,
+                                                         functor);
     }
   }
 };
