+++ paddle/fluid/operators/elementwise/elementwise_op_broadcast.cu.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "paddle/fluid/operators/elementwise/elementwise_op_impl.cu.h"
+#include "paddle/phi/kernels/funcs/elementwise_op_impl.cu.h"
 
 // only can include the headers in paddle/top/api dirs
 #include "paddle/phi/kernels/gpu/elementwise.h"
