+++ paddle/phi/kernels/cpu/unfold_grad_kernel.cc
@@ -13,9 +13,10 @@
 // limitations under the License.
 
 #include "paddle/phi/kernels/unfold_grad_kernel.h"
+#include "paddle/phi/kernels/impl/unfold_grad_kernel_impl.h"
+
 #include "paddle/phi/backends/cpu/cpu_context.h"
 #include "paddle/phi/core/kernel_registry.h"
-#include "paddle/phi/kernels/impl/unfold_grad_kernel_impl.h"
 
 PD_REGISTER_KERNEL(
     unfold_grad, CPU, ALL_LAYOUT, phi::UnfoldGradKernel, float, double) {}
