+++ paddle/phi/kernels/funcs/elementwise_op_impl.cu.h
@@ -13,32 +13,31 @@ See the License for the specific language governing permissions and
 limitations under the License. */
 
 #pragma once
+#include <memory>
 
-#include "paddle/fluid/framework/pten_utils.h"
-#include "paddle/fluid/framework/tensor.h"
-
-// only can include the headers in paddle/top/api dirs
 #include "paddle/phi/api/lib/utils/tensor_utils.h"
+#include "paddle/phi/core/dense_tensor.h"
 #include "paddle/phi/kernels/funcs/elementwise_base.h"
+#include "paddle/phi/kernels/primitive/kernel_primitives.h"
 
-namespace paddle {
-namespace operators {
-
-using ElementwiseType = phi::ElementwiseType;
+namespace phi {
+namespace funcs {
 
 template <typename OutT, typename Functor, int NumOuts = 1>
 void LaunchSameDimsElementwiseCudaKernel(
-    const KPDevice &ctx, const std::vector<const framework::Tensor *> &ins,
-    std::vector<framework::Tensor *> *outs, Functor func) {
-  std::vector<const phi::DenseTensor *> pt_inputs;
-  std::vector<phi::DenseTensor *> pt_outputs;
+    const KPDevice &ctx,
+    const std::vector<const DenseTensor *> &ins,
+    std::vector<DenseTensor *> *outs,
+    Functor func) {
+  std::vector<const DenseTensor *> pt_inputs;
+  std::vector<DenseTensor *> pt_outputs;
   // TODO(YuanRisheng) *_tmp for cache DenseTensor, because the temporary
   // DenseTensor obj
   // generated by MakePtenDenseTensor can be destroyed when exits loop. *_tmp
   // can be deleted
   // when DenseTensor support copy constructor.
-  std::vector<std::unique_ptr<phi::DenseTensor>> pt_inputs_tmp;
-  std::vector<std::unique_ptr<phi::DenseTensor>> pt_outputs_tmp;
+  std::vector<std::unique_ptr<DenseTensor>> pt_inputs_tmp;
+  std::vector<std::unique_ptr<DenseTensor>> pt_outputs_tmp;
   for (auto in : ins) {
     pt_inputs_tmp.emplace_back(
         std::move(paddle::experimental::MakePtenDenseTensor(*in)));
@@ -53,9 +52,9 @@ void LaunchSameDimsElementwiseCudaKernel(
   for (int i = 0; i < pt_outputs_tmp.size(); i++) {
     pt_outputs.push_back(pt_outputs_tmp[i].get());
   }
-  phi::funcs::ElementwiseKernel<OutT, Functor, NumOuts>(ctx, pt_inputs,
-                                                        &pt_outputs, func);
+  phi::funcs::ElementwiseKernel<OutT, Functor, NumOuts>(
+      ctx, pt_inputs, &pt_outputs, func);
 }
 
-}  // namespace operators
-}  // namespace paddle
+}  // namespace funcs
+}  // namespace phi
