+++ paddle/phi/kernels/gpu/label_smooth_grad_kernel.cu
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/operators/elementwise/elementwise_op_impl.cu.h"
+#include "paddle/phi/kernels/label_smooth_grad_kernel.h"
+
 #include "paddle/phi/backends/gpu/gpu_context.h"
 #include "paddle/phi/core/kernel_registry.h"
-#include "paddle/phi/kernels/label_smooth_grad_kernel.h"
+#include "paddle/phi/kernels/funcs/elementwise_op_impl.cu.h"
 
 namespace phi {
 template <typename T>
@@ -41,7 +42,7 @@ void LabelSmoothGradKernel(const Context& dev_ctx,
   std::vector<const DenseTensor*> ins = {&out_grad};
   std::vector<DenseTensor*> outs = {label_grad};
   auto functor = LabelSmoothGradFunctor<T>(epsilon);
-  paddle::operators::LaunchSameDimsElementwiseCudaKernel<T>(
+  phi::funcs::LaunchSameDimsElementwiseCudaKernel<T>(
       dev_ctx, ins, &outs, functor);
 }
