+++ paddle/phi/kernels/gpu/unfold_grad_kernel.cu
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "paddle/phi/kernels/unfold_grad_kernel.h"
+
 #include "paddle/phi/backends/gpu/gpu_context.h"
 #include "paddle/phi/core/kernel_registry.h"
 #include "paddle/phi/kernels/impl/unfold_grad_kernel_impl.h"
-#include "paddle/phi/kernels/unfold_grad_kernel.h"
 
 PD_REGISTER_KERNEL(
     unfold_grad, GPU, ALL_LAYOUT, phi::UnfoldGradKernel, float, double) {}
